extract reusable test code to separate module
