+++ build.gradle
@@ -24,7 +24,7 @@ allprojects {
 }
 
 subprojects {
-    if(it.parent.name == 'examples' || it.parent.name == 'test') {
+    if(it.parent.name == 'examples') {
         apply plugin: 'java'
     } else {
         apply plugin: 'java-library'
