+++ settings.gradle
@@ -13,3 +13,5 @@ include(':metrics-reporter:metrics-reporter-datadog-spring-boot')
 include(':examples:examples-common')
 include(':examples:example-spring-boot-starter-web')
 include(':examples:example-spring-boot-starter-webflux')
+
+include(':test:test-utils')
