+++ test/test-utils/src/main/resources/openapi.yaml
